fix(endpoint): move function to own file
