+++ packages/middleware-endpoint/src/adaptors/getEndpointFromInstructions.ts
@@ -1,8 +1,9 @@
-import { Endpoint, EndpointParameters, EndpointV2, HandlerExecutionContext } from "@aws-sdk/types";
+import { EndpointParameters, EndpointV2, HandlerExecutionContext } from "@aws-sdk/types";
 
 import { EndpointResolvedConfig } from "../resolveEndpointConfig";
 import { resolveParamsForS3 } from "../service-customizations";
 import { EndpointParameterInstructions } from "../types";
+import { createConfigValueProvider } from "./createConfigValueProvider";
 
 export type EndpointParameterInstructionsSupplier = Partial<{
   getEndpointParameterInstructions(): EndpointParameterInstructions;
@@ -64,7 +65,7 @@ export const resolveParams = async <
         break;
       case "clientContextParams":
       case "builtInParams":
-        endpointParams[name] = await createConfigProvider<Config>(instruction.name, name, clientConfig)();
+        endpointParams[name] = await createConfigValueProvider<Config>(instruction.name, name, clientConfig)();
         break;
       default:
         throw new Error("Unrecognized endpoint parameter instruction: " + JSON.stringify(instruction));
@@ -80,47 +81,4 @@ export const resolveParams = async <
   }
 
   return endpointParams;
-};
-
-/**
- * Normalize some key of the client config to an async provider.
- * @private
- *
- * @param configKey - the key to look up in config.
- * @param canonicalEndpointParamKey - this is the name the EndpointRuleSet uses.
- *                                    it will most likely not contain the config
- *                                    value, but we use it as a fallback.
- * @param config - container of the config values.
- * 
- * @returns async function that will resolve with the value.
- */
-export const createConfigProvider = <Config extends Record<string, unknown>>(
-  configKey: string,
-  canonicalEndpointParamKey: string,
-  config: Config
-) => {
-  const configProvider = async () => {
-    const configValue: unknown = config[configKey] ?? config[canonicalEndpointParamKey];
-    if (typeof configValue === "function") {
-      return configValue();
-    }
-    return configValue;
-  };
-  if (configKey === "endpoint" || canonicalEndpointParamKey === "endpoint") {
-    return async () => {
-      const endpoint = await configProvider();
-      if (endpoint && typeof endpoint === "object") {
-        if ("url" in endpoint) {
-          return (endpoint as EndpointV2).url.href;
-        }
-        if ("hostname" in endpoint) {
-          const { protocol, hostname, port, path } = endpoint as Endpoint;
-          // query params are ignored in setting endpoint.
-          return `${protocol}//${hostname}${port ? ":" + port : ""}${path}`;
-        }
-      }
-      return endpoint;
-    };
-  }
-  return configProvider;
 };
